refactor(codegen): decompose states check by concern
