+++ scripts/codegen/src/plugins/states/check/transitions.ts
@@ -1,141 +1,15 @@
-import type { Issue } from "../../core/check-utils.ts";
-import { isComposite, issue, suggestionFragment } from "../../core/check-utils.ts";
-import type { Vocabulary } from "../../lib/vocabulary.ts";
-import type { PayloadEntry, Spec, SpecPart } from "../../schema.ts";
+import type { Issue } from "../../../core/check-utils.ts";
+import { issue, suggestionFragment } from "../../../core/check-utils.ts";
+import type { PayloadEntry, SpecPart } from "../../../schema.ts";
+import type { PartRef } from "./part-names.ts";
 
-/**
- * Rejections for per-part `states:` blocks:
- *
- *  - Empty `states:` rejected; initial state = first key.
- *  - Transition `to:` targets resolve in the same part's `states:` map.
- *  - Source-key prefix resolves: `<part>.<event>` → part exists in the
- *    parts tree and `<event>` is a registered DOM event;
- *    `key.<name>` → `<name>` is a registered key name;
- *    `outside.<event>` → current part declares `overlay:`.
- *  - Part-name uniqueness across the spec's parts tree (so `<part>.<event>`
- *    references stay unambiguous without a qualified path).
- *  - `emits:` event names exist in root `events:`; payload literals match
- *    the declared payload shape (closed enums checked exactly).
- *  - `overlay.anchor` resolves to a sibling part declaring `fromChildren: true`.
- *  - `pattern: "controllable"` props mirror a state name when the part
- *    also declares `states:`.
- *  - `when:` guards parse as `[!]<part>.<bool-prop>`; part exists, prop is a
- *    `type: boolean` declared on that part.
- *  - `after:` references a declared `type: number` prop on the same part —
- *    the generator emits `setTimeout(..., props.<name>)` which fails
- *    silently if the prop is the wrong type.
- *
- * Unreachable-event detection (an event declared in root `events:` but never
- * fired from any transition's `emits:`) is intentionally not enforced: the
- * Issue type has no severity axis today, and "warn but allow" needs that
- * infra to land first.
- */
-export function checkStateMachines(spec: Spec, vocabulary: Vocabulary): Issue[] {
-  if (!isComposite(spec)) return [];
-  const issues: Issue[] = [];
-
-  type PartRef = { part: SpecPart; path: string };
-  const partsByName = new Map<string, PartRef>();
-  const partNameOccurrences = new Map<string, string[]>();
-  const walkParts = (parts: Record<string, SpecPart>, basePath: string): void => {
-    for (const [name, part] of Object.entries(parts)) {
-      const path = basePath === "" ? `parts.${name}` : `${basePath}.parts.${name}`;
-      const seen = partNameOccurrences.get(name) ?? [];
-      seen.push(path);
-      partNameOccurrences.set(name, seen);
-      if (!partsByName.has(name)) partsByName.set(name, { part, path });
-      if (part.parts) walkParts(part.parts, path);
-    }
-  };
-  walkParts(spec.parts, "");
-
-  // Rule 3a — part name uniqueness across the parts tree.
-  for (const [name, paths] of partNameOccurrences) {
-    if (paths.length <= 1) continue;
-    issues.push(
-      issue(
-        spec.name,
-        paths[0] ?? "parts",
-        `part name '${name}' is declared ${paths.length} times (${paths.join(", ")}). Part names must be unique across the parts tree so '<part>.<event>' source references stay unambiguous.`,
-      ),
-    );
-  }
-
-  const domEvents = vocabulary.dom_events ?? {};
-  const keys = vocabulary.keys ?? {};
-  const domEventNames = Object.keys(domEvents);
-  const keyNames = Object.keys(keys);
-
-  const declaredEvents = spec.events ?? {};
-  const declaredEventNames = Object.keys(declaredEvents);
-
-  // Rule 6 — overlay.anchor must point at a sibling part that declares
-  // `fromChildren: true`. Fires whether or not the part declares `states:`.
-  const visitOverlay = (
-    parts: Record<string, SpecPart>,
-    siblings: Record<string, SpecPart>,
-    basePath: string,
-  ): void => {
-    for (const [name, part] of Object.entries(parts)) {
-      const path = basePath === "" ? `parts.${name}` : `${basePath}.parts.${name}`;
-      if (part.overlay) {
-        const anchorName = part.overlay.anchor;
-        const sibling = siblings[anchorName];
-        if (!sibling) {
-          issues.push(
-            issue(
-              spec.name,
-              `${path}.overlay.anchor`,
-              `'${anchorName}' is not a sibling part of '${name}'. \`overlay.anchor\` must name a part declared alongside this one.`,
-            ),
-          );
-        } else if (sibling.fromChildren !== true) {
-          issues.push(
-            issue(
-              spec.name,
-              `${path}.overlay.anchor`,
-              `'${anchorName}' must declare \`fromChildren: true\` to serve as an overlay anchor. The anchor wraps the consumer's children and supplies the anchor element.`,
-            ),
-          );
-        }
-      }
-      if (part.parts) visitOverlay(part.parts, part.parts, path);
-    }
-  };
-  visitOverlay(spec.parts, spec.parts, "");
-
-  // Rules 1, 2, 3, 5, 7 (via 3), 9, 10, +after — walk every part that
-  // declares states:.
-  const visitStates = (parts: Record<string, SpecPart>, basePath: string): void => {
-    for (const [partName, part] of Object.entries(parts)) {
-      const partPath = basePath === "" ? `parts.${partName}` : `${basePath}.parts.${partName}`;
-      if (part.states) {
-        checkPartStates(spec.name, partName, part, partPath, {
-          partsByName,
-          domEventNames,
-          domEvents,
-          keyNames,
-          keys,
-          declaredEvents,
-          declaredEventNames,
-          issues,
-        });
-      }
-      if (part.parts) visitStates(part.parts, partPath);
-    }
-  };
-  visitStates(spec.parts, "");
-
-  return issues;
-}
-
-function checkPartStates(
+export function checkPartStates(
   specName: string,
   partName: string,
   part: SpecPart,
   partPath: string,
   ctx: {
-    partsByName: Map<string, { part: SpecPart; path: string }>;
+    partsByName: Map<string, PartRef>;
     domEventNames: string[];
     domEvents: Record<string, string>;
     keyNames: string[];
