+++ scripts/codegen/src/semantic-checks.ts
@@ -23,7 +23,7 @@ import { checkMotion } from "./plugins/motion/check.ts";
 import { checkRepeatingParts } from "./plugins/parts/check/index.ts";
 import { checkResponsiveExplicit } from "./plugins/props/check.ts";
 import { checkRoot } from "./plugins/root/check.ts";
-import { checkStateMachines } from "./plugins/states/check.ts";
+import { checkStateMachines } from "./plugins/states/check/index.ts";
 import {
   checkCssImportAllowlist,
   checkPrivateTokens,
@@ -102,7 +102,7 @@ export { checkImperativeProps } from "./plugins/imperativeProps/check.ts";
 export { checkRepeatingParts } from "./plugins/parts/check/index.ts";
 export { checkResponsiveExplicit } from "./plugins/props/check.ts";
 export { checkRoot } from "./plugins/root/check.ts";
-export { checkStateMachines } from "./plugins/states/check.ts";
+export { checkStateMachines } from "./plugins/states/check/index.ts";
 export {
   checkCssImportAllowlist,
   checkPrivateTokens,
